check số lượng hàng khi đặt

+++ bookstore/src/main/resources/static/js/customer/order.controller.js
@@ -18,10 +18,14 @@
         vm.remove = remove;
 
         function remove(item) {
-            let idx = vm.cartList.indexOf(item);
-            vm.cartList.splice(idx, 1);
-            console.log(vm.cartList)
-            resetCart();
+            AlertService.showConfirm("Xác nhận xóa sản phẩm").then(function () {
+                let idx = vm.cartList.indexOf(item);
+                vm.cartList.splice(idx, 1);
+                console.log(vm.cartList)
+                resetCart();
+            }, function () {
+
+            })
         }
 
         vm.createBuyDto = {
@@ -50,19 +54,68 @@
 
         vm.cartList = [];
         getCart();
-        function getCart(){
+
+        function getCart() {
             let storage = localStorage.getItem("cartList");
             if (storage != null) {
                 vm.cartList = JSON.parse(storage);
+                vm.cartList.forEach(function (e) {
+                    $.ajax({
+                        url: "/api/bookQuantity/getOneByStore",
+                        type: "GET",
+                        data: {
+                            bookId: e.bookId,
+                            storeId: e.storeId
+                        },
+                        success: function (response) {
+                            $scope.$apply(function () {
+                                if (response.errorCode == 0) {
+                                    e.amount = response.data.quantity;
+                                }
+                            })
+                        }
+                    })
+                })
                 calculate();
             }
+        }
+
+        vm.changeStore = changeStore;
+
+        function changeStore(item) {
+            $.ajax({
+                url: "/api/bookQuantity/getOneByStore",
+                type: "GET",
+                data: {
+                    bookId: item.bookId,
+                    storeId: item.storeId
+                },
+                success: function (response) {
+                    $scope.$apply(function () {
+                        if (response.errorCode == 0) {
+                            item.amount = response.data.quantity;
+                        }
+                    })
+                }
+            })
         }
 
         function order() {
             vm.createBuyDto.buyBookDtoList = vm.cartList;
             vm.createBuyDto.totalMoney = vm.totalMoney;
             console.log(vm.createBuyDto);
-            if (vm.totalMoney == 0 || isNaN(vm.totalMoney)){
+            let oversize = false;
+            vm.createBuyDto.buyBookDtoList.forEach(function (e) {
+                if (e.amount < e.quantity) {
+                    oversize = true;
+
+                }
+            })
+            if (oversize) {
+                AlertService.error("Số lượng còn lại không đủ, xin kiểm tra lại giỏ hàng.", 2000);
+                return;
+            }
+            if (vm.totalMoney == 0 || isNaN(vm.totalMoney)) {
                 AlertService.error("Bạn phải chọn tối thiểu 1 sản phẩm để thanh toán", 2000, "bottom right");
                 return;
             }
@@ -76,11 +129,16 @@
                 contentType: "application/json",
                 data: JSON.stringify(vm.createBuyDto),
                 success: function (response) {
-                    AlertService.success("Đặt hàng thành công!", 2000);
-                    localStorage.clear();
-                    setTimeout(function () {
-                        location.href = "/order/success/" + response.data.id;
-                    }, 2000)
+                    if (response.errorCode == 0) {
+                        AlertService.success("Đặt hàng thành công", 2000);
+                        localStorage.clear();
+                        setTimeout(function () {
+                            location.href = "/order/success/" + response.data.id;
+                        }, 2000)
+                    } else {
+                        AlertService.error("Lỗi", 2000);
+                    }
+
                 }
             })
         }
